Complete shell ownership quality repair
