+++ apps/mobile/src/lib/shell-controllers/scrollback-core.ts
@@ -19,6 +19,7 @@ import {
 	type ControllerInvalidationReason,
 	type ControllerOutcome,
 } from './controller-core';
+import { matchControllerOutcome } from './controller-outcome';
 import { handleScrollbackBatch } from './scrollback-batch-coordinator';
 import {
 	createSafeScrollbackWarn as createSafeWarn,
@@ -295,8 +296,43 @@ export function createShellScrollbackControllerCore(
 			executor === createdExecutor &&
 			executorRevision === capturedExecutorRevision;
 		try {
+			const toExecutorResult = async (
+				operation: Promise<
+					Awaited<ReturnType<typeof nextContext.workmux.scroll.enter>>
+				>,
+			) => {
+				const outcome = await operation;
+				return matchControllerOutcome(outcome, {
+					completed: (completed) => ({
+						success: true,
+						output: completed.output ?? '',
+					}),
+					failed: (failed) => ({
+						success: false,
+						output: failed.output ?? '',
+						error: failed.failure.message,
+					}),
+					superseded: () => ({
+						success: false,
+						output: '',
+						error: 'Workmux scroll command superseded.',
+					}),
+					unavailable: () => ({
+						success: false,
+						output: '',
+						error: 'Workmux scroll command unavailable.',
+					}),
+				});
+			};
 			createdExecutor = createExecutor({
-				scrollTransport: nextContext.workmux.scroll,
+				scrollTransport: {
+					enter: (input) =>
+						toExecutorResult(nextContext.workmux.scroll.enter(input)),
+					move: (input) =>
+						toExecutorResult(nextContext.workmux.scroll.move(input)),
+					exit: (input) =>
+						toExecutorResult(nextContext.workmux.scroll.exit(input)),
+				},
 				onFailure: (message, failureContext) => {
 					if (!isCurrentExecutor()) return;
 					const currentContext = context;
