+++ apps/mobile/src/lib/shell-controllers/terminal-contracts.ts
@@ -1,3 +1,14 @@
-import { type ShellTerminalRuntimeView } from './terminal-hook-runtime';
+import { type TerminalRuntimeKey } from './terminal-transport';
 
-export type ShellTerminalViewPort = ShellTerminalRuntimeView;
+export type ShellTerminalViewPort = {
+	getRuntimeKey(): TerminalRuntimeKey | null;
+	getRuntimeInstanceId(): string | null;
+	getSelectionModeEnabled(): boolean;
+	isCurrentInstance(instanceId: string): boolean;
+	fit(): void;
+	setSystemKeyboardEnabled(enabled: boolean): void;
+	setSelectionModeEnabled(enabled: boolean): void;
+	getSelection(): Promise<string>;
+	exitScrollback(message: { requestId: number; instanceId?: string }): void;
+	sendScrollbackEnterAck(requestId: number, instanceId: string): void;
+};
